requirements.txt
# 필수 패키지 (웹 프레임워크)
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
pydantic>=1.8.0
python-dotenv>=0.19.0

# 이미지 처리
pillow>=9.0.0
opencv-python>=4.5.0
ultralytics>=8.0.0

# 데이터 처리
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.0.0

# 텍스트 요약 (TextRank)
summa>=1.2.0

# HTTP 요청
requests>=2.25.0

# 지도 시각화
folium>=0.12.0

# 챗봇 기능 (선택적 - 설치 실패해도 앱 작동)
langchain==0.3.13
langchain-core==0.3.29
langchain-community==0.3.13
langchain-google-genai==2.0.5
chromadb==0.5.15
google-generativeai==0.8.5
huggingface-hub==0.36.0
sentence-transformers==3.0.1
python-dotenv==1.2.1
jinja2==3.1.6
Shapely
